Switching from double to long double

+++ mmmalloc/helpers/c_dynamic_priority_queue/element.h
@@ -8,7 +8,7 @@
 #include <cmath>
 
 typedef unsigned dpq_name_t;
-typedef double dpq_time_t;
+typedef long double dpq_time_t;
 
 
 class Element {
@@ -22,37 +22,37 @@ class Element {
   bool operator<(const Element& rhs) const;
   bool operator==(const Element& rhs) const;
   void update(dpq_time_t current_time) const;
-  double get_switch_time(const Element& other_element) const;
+  long double get_switch_time(const Element& other_element) const;
   dpq_name_t get_name() const;
-  double get_cpu_credibility() const;
-  double get_memory_credibility() const;
-  double get_priority() const;
+  long double get_cpu_credibility() const;
+  long double get_memory_credibility() const;
+  long double get_priority() const;
   dpq_time_t get_update_time() const;
-  double get_cpu_relative_allocation() const;
-  double get_memory_relative_allocation() const;
-  void set_cpu_relative_allocation(double cpu_relative_allocation);
-  void set_memory_relative_allocation(double memory_relative_allocation);
+  long double get_cpu_relative_allocation() const;
+  long double get_memory_relative_allocation() const;
+  void set_cpu_relative_allocation(long double cpu_relative_allocation);
+  void set_memory_relative_allocation(long double memory_relative_allocation);
   operator std::string() const;
 
  private:
   struct Resource {
-    Resource(double system_total, double credibility=0,
-             double relative_allocation=0);
-    const double system_total;
-    mutable double credibility;
-    double relative_allocation;
+    Resource(long double system_total, long double credibility=0,
+             long double relative_allocation=0);
+    const long double system_total;
+    mutable long double credibility;
+    long double relative_allocation;
   };
   mutable dpq_time_t update_time;
-  const double tau;
+  const long double tau;
   Resource cpu;
   Resource memory;
 
-  double calculate_credibility(dpq_time_t current_time,
-    double previous_credibility, double relative_allocation) const;
+  long double calculate_credibility(dpq_time_t current_time,
+    long double previous_credibility, long double relative_allocation) const;
   dpq_time_t get_priority_intersection(const Resource& r1,
                                        const Resource& r2) const;
-  double calculate_priority(const Resource& res, const dpq_time_t time=0) const;
-  double get_priority_derivative(const Resource& r, dpq_time_t time_delta) const;
+  long double calculate_priority(const Resource& res, const dpq_time_t time=0) const;
+  long double get_priority_derivative(const Resource& r, dpq_time_t time_delta) const;
   const Resource& get_dominant_resource(const dpq_time_t time=0) const;
 };
